🎓 Added optional ML emote classifier
To tackle the problem of having inflexible and probably invalid data for the emote analysis I have tested a simple machine learning algorithm for emote classification. I also included a webscraper that pulls the top used emotes from stream elements automatically. The classifier still has a lot of problems that need to be fixed and maybe my current approach doesn't even work, but I feel like it is a step in the right direction to improve Moodmonitor.