🏗️ Major refactoring
Refactored code to make it more manageable and to avoid global variables. Also the script  has been split in multiple modules for the same purpose. The collected data is now added instead of overwritten. This makes it possible to run multiple scripts at once and collect data in the same json file.